Respond to PR feedback
- Remove unused model hook
- Extract ComponentViewItem to a separate file
